refactor: create 2 separate signup-discount-notice components for A/B testing
