+++ app/components/pay-page/signup-discount-notice.hbs
@@ -1,34 +1,13 @@
-{{#if (eq @variant "default")}}
-  <div class="flex p-3 bg-blue-100 border border-blue-300 rounded shadow-sm relative" data-test-signup-discount-notice ...attributes>
-    <div class="mr-3 mt-0.5">
-      {{svg-jar "clock" class="w-6 fill-current text-blue-400 animate-pulse"}}
-    </div>
-
-    <div class="text-blue-700 prose prose-blue leading-7">
-      Limited offer: Subscribe by
-      <span class="font-semibold percy-timestamp">{{date-format @discount.expiresAt format="PPPp"}}</span>
-      to get
-      <span class="font-semibold">40% off</span>
-      the annual plan.
-    </div>
+<div class="flex p-3 bg-blue-100 border border-blue-300 rounded shadow-sm relative" data-test-signup-discount-notice ...attributes>
+  <div class="mr-3 mt-0.5">
+    {{svg-jar "clock" class="w-6 fill-current text-blue-400 animate-pulse"}}
   </div>
-{{else}}
-  <div
-    class="flex p-3.5 bg-gradient-to-b from-gray-600 to-gray-900 border border-yellow-600 rounded relative shadow"
-    data-test-signup-discount-notice
-    ...attributes
-  >
-    <div class="mr-1.5 flex-shrink-0">
-      <img src={{this.iconImage}} alt="treasure" class="w-8 animate-spin-once" />
-    </div>
 
-    <div class="bg-repeating-gradient-to-b from-yellow-100 to-amber-500 text-transparent bg-clip-text leading-7 inline-block mt-0.5">
-      New user offer: Subscribe in
-      <span class="font-bold percy-timestamp font-mono text-lg border-b border-yellow-500 border-dashed">{{this.timeLeftText}}</span>
-      to get
-      <span class="font-bold">40% off</span>
-      the annual plan.
-    </div>
-    <EmberTooltip @text="The 1 year plan is usually $360, but you can get it for $216 with this offer." @side="bottom" />
+  <div class="text-blue-700 prose prose-blue leading-7">
+    Limited offer: Subscribe by
+    <span class="font-semibold percy-timestamp">{{date-format @discount.expiresAt format="PPPp"}}</span>
+    to get
+    <span class="font-semibold">40% off</span>
+    the annual plan.
   </div>
-{{/if}}
+</div>
