+++ app/components/pay-page/signup-discount-notice.ts
@@ -1,27 +1,15 @@
 import Component from '@glimmer/component';
 import PromotionalDiscountModel from 'codecrafters-frontend/models/promotional-discount';
-import iconImage from '/assets/images/icons/money.svg';
-import { service } from '@ember/service';
-import type TimeService from 'codecrafters-frontend/services/time';
-import { formatTimeLeft } from 'codecrafters-frontend/utils/time-formatting';
+
 interface Signature {
   Element: HTMLDivElement;
 
   Args: {
     discount: PromotionalDiscountModel;
-    variant?: 'default' | 'countdown';
   };
 }
 
-export default class SignupDiscountNoticeComponent extends Component<Signature> {
-  iconImage = iconImage;
-
-  @service declare time: TimeService;
-
-  get timeLeftText() {
-    return formatTimeLeft(this.args.discount.expiresAt, this.time.currentTime);
-  }
-}
+export default class SignupDiscountNoticeComponent extends Component<Signature> {}
 
 declare module '@glint/environment-ember-loose/registry' {
   export default interface Registry {
